requirements.txt
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.10.0
babel==2.17.0
billiard==4.2.3
brotli==1.2.0
celery==5.5.3
cffi==2.0.0
cfgv==3.4.0
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cssselect2==0.8.0
distlib==0.4.0
Django==5.2.7
django-ajax-datatable==4.5.0
django-environ==0.12.0
django-hCaptcha==0.2.0
django-ninja==1.4.5
django-ranged-response==0.2.0
et_xmlfile==2.0.0
filelock==3.20.0
fonttools==4.60.1
identify==2.6.15
kombu==5.5.4
nodeenv==1.9.1
openpyxl==3.1.5
packaging==25.0
pillow==11.3.0
platformdirs==4.5.0
pre_commit==4.3.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
pydyf==0.11.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
redis==7.1.0
six==1.17.0
sqlparse==0.5.3
tinycss2==1.5.1
tinyhtml5==2.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
vine==5.1.0
virtualenv==20.35.3
wcwidth==0.2.14
weasyprint==66.0
webencodings==0.5.1
xlsx2html==0.6.2
zopfli==0.4.0
